services/embed_indexing/requirements.txt
fastapi
uvicorn[standard]
openai
faiss-cpu
numpy
pip install sentence-transformers
